Implement linting job when committing code to main.
Renamed test.yaml to source_phase_linting.yaml. This is meant for a GitHub action that is automatically run when merging into main.

If the basic linting test fails, the pull request cannot be merged until the issue is resolved and the lint job passes. 

Once the lint job passes, the pull request becomes eligible for merging into main. 

This source phase lint job is part of the CI/CD pipeline