Added warning light script
The warning_light.py script uses CircuitPython libraries and the
Neopixel library to control a WS2812B LED strip connected to the
Raspberry Pi (on 5V, GND and GPIO-18). LED strip configuration
variables are used to initialize the LED strip, and then blink the
LED strip on and off every second, for a period of 10 seconds.

Final changes to this script would involve the configuration and
creation of the LED strip being wrapped in a function, and simply
calling that function when a network request requires it to do so.