Created server.py, which uses Python's htttp.server package to send and receive http responses
A handler class called ServerHandler was created, which has a handle request
method. This method takes in an HTTP method, and parses the request path.
A route decorator pattern is used to make it easier to add multiple handlers
to the ServerHandler. The route decorator takes in an HTTP method and absolute
route.

A route, with an HTTP method of post and a path of /warning light
is set for the warning light handler, which gets a JSON payload
from the POST request and parses this payload. This payload should
receive three fields: color, which is a list of three integers (RGB),
duration, how long the warning lights should flicker and rate, the
rate at which the warning lights should flicker. These are validated
and the appropriate response is sent.